pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medallion"
version = "1.0.0"
description = "A Python-first agentic AI framework with Go backend"
authors = [{name = "Medallion Team", email = "team@medallion.ai"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["ai", "agents", "workflow", "knowledge-graph", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
medallion = "medallion.cli:main"

[project.urls]
Homepage = "https://github.com/medallion-ai/medallion"
Repository = "https://github.com/medallion-ai/medallion"
Documentation = "https://medallion.readthedocs.io"
Issues = "https://github.com/medallion-ai/medallion/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["medallion*"]

[tool.setuptools.package-data]
medallion = ["bin/*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
